backend/requirements.txt
opencv-python==4.10.0.84
mediapipe==0.10.15
numpy==1.26.4
Flask==2.3.2
Flask-CORS==3.0.10
requests==2.28.2
python-dotenv==0.19.2
openai==1.0.0
websockets==10.4
sounddevice==0.5.0